Moving things around between organizations, funders, journals
