+++ dspace/config/dspace.cfg
@@ -2618,6 +2618,8 @@ solrauthority.searchfieldtype = full-text
 solrauthority.searchscheme.prism_publicationName = Journal
 solrauthority.sortfieldtype = display-value
 solrauthority.searchscheme.dc_contributor_author = Author
+solrauthority.searchscheme.dryad_fundingEntity = Funder
+solrauthority.searchscheme.dryad_organization = Organization
 
 core.authorization.site-admin.group=Senior Curators
