+++ dspace/modules/api/src/main/java/org/datadryad/api/DryadFunderConcept.java
@@ -14,35 +14,23 @@ import java.lang.*;
 import java.lang.Exception;
 import java.sql.SQLException;
 
+import static org.datadryad.api.DryadJournalConcept.PAYMENT_PLAN;
+import static org.datadryad.api.DryadJournalConcept.WEBSITE;
+
 /**
  *
  * @author Daisie Huang <daisieh@datadryad.org>
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class DryadFunderConcept extends DryadOrganizationConcept {
-    // Journal Concepts can have the following metadata properties defined in the old properties file:
-    public static final String FULLNAME = "fullname";
-    public static final String PAYMENT_PLAN = "paymentPlanType";
-
-    // Journal Concepts can also have the following fields:
-    public static final String CUSTOMER_ID = "customerID";
-    public static final String DESCRIPTION = "description";
-    public static final String WEBSITE = "website";
+    public static final String FUNDER_ID = "identifier";
 
     private static Logger log = Logger.getLogger(DryadFunderConcept.class);
 
     static {
-        metadataProperties.setProperty(FULLNAME, "journal.fullname");
-        metadataProperties.setProperty(PAYMENT_PLAN, "journal.paymentPlanType");
-        metadataProperties.setProperty(CUSTOMER_ID, "journal.customerID");
-        metadataProperties.setProperty(DESCRIPTION, "journal.description");
-        metadataProperties.setProperty(WEBSITE, "journal.website");
+        metadataProperties.setProperty(FUNDER_ID, "funder.identifier");
 
-        defaultMetadataValues.setProperty(metadataProperties.getProperty(FULLNAME), "");
-        defaultMetadataValues.setProperty(metadataProperties.getProperty(PAYMENT_PLAN), "");
-        defaultMetadataValues.setProperty(metadataProperties.getProperty(CUSTOMER_ID), "");
-        defaultMetadataValues.setProperty(metadataProperties.getProperty(DESCRIPTION), "");
-        defaultMetadataValues.setProperty(metadataProperties.getProperty(WEBSITE), "");
+        defaultMetadataValues.setProperty(metadataProperties.getProperty(FUNDER_ID), "");
     }
 
     // these are mandatory elements
@@ -80,66 +68,18 @@ public class DryadFunderConcept extends DryadOrganizationConcept {
         }
     }
 
+    @Override
     public void create(Context context) {
         try {
             context.turnOffAuthorisationSystem();
-            Scheme journalScheme = Scheme.findByIdentifier(context, ConfigurationManager.getProperty("solrauthority.searchscheme.prism_publicationName"));
-            Concept newConcept = journalScheme.createConcept(context);
+            Scheme funderScheme = Scheme.findByIdentifier(context, ConfigurationManager.getProperty("solrauthority.searchscheme.dryad_fundingEntity"));
+            Concept newConcept = funderScheme.createConcept(context);
             this.setUnderlyingConcept(context, newConcept);
             context.commit();
             context.restoreAuthSystemState();
         } catch (Exception e) {
             log.error("Couldn't make new concept: " + e.getMessage());
         }
-    }
-
-    public void delete(Context context) throws SQLException, AuthorizeException {
-        this.getUnderlyingConcept(context).delete(context);
-    }
-
-    public String getCustomerID() {
-        return getConceptMetadataValue(metadataProperties.getProperty(CUSTOMER_ID));
-    }
-
-    public void setCustomerID(String value) {
-        setConceptMetadataValue(metadataProperties.getProperty(CUSTOMER_ID), value);
-    }
-
-    public String getDescription() {
-        return getConceptMetadataValue(metadataProperties.getProperty(DESCRIPTION));
-    }
-
-    public void setDescription(String value) {
-        setConceptMetadataValue(metadataProperties.getProperty(DESCRIPTION), value);
-    }
-
-    public String getWebsite() {
-        return getConceptMetadataValue(metadataProperties.getProperty(WEBSITE));
-    }
-
-    public void setWebsite(String value) {
-        setConceptMetadataValue(metadataProperties.getProperty(WEBSITE), value);
-    }
-
-    public String getPaymentPlan() {
-        return getConceptMetadataValue(metadataProperties.getProperty(PAYMENT_PLAN));
-    }
-
-    public void setPaymentPlan(String paymentPlan) {
-        String paymentPlanType = "";
-        if (SUBSCRIPTION_PLAN.equals(paymentPlan)) {
-            paymentPlanType = SUBSCRIPTION_PLAN;
-        } else if (DEFERRED_PLAN.equals(paymentPlan)) {
-            paymentPlanType = DEFERRED_PLAN;
-        } else if (PREPAID_PLAN.equals(paymentPlan)) {
-            paymentPlanType = PREPAID_PLAN;
-        }
-        setConceptMetadataValue(metadataProperties.getProperty(PAYMENT_PLAN), paymentPlanType);
-    }
-
-    @JsonIgnore
-    public Boolean isValid() {
-        return (getFullName() != null);
     }
 
     @JsonIgnore
